Get rid of circ. deps. Move containerTitlePipe into toString
