+++ src/app/metadata/metadata.component.ts
@@ -8,7 +8,7 @@ import {
 import { LocdbService } from '../locdb.service';
 import { Component, Input, OnInit, OnChanges } from '@angular/core';
 
-import { AuthorsPipe, EditorsPipe, PublisherPipe, EmbracePipe, PrefixPipe, ContainerTitlePipe } from '../pipes';
+import { AuthorsPipe, EditorsPipe, PublisherPipe, EmbracePipe, PrefixPipe } from '../pipes';
 
 import { StandardPipe, ContainerPipe } from '../pipes';
