--- src/app/pipes/container-title.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { ContainerTitlePipe } from './container-title.pipe';
-
-describe('ContainerTitlePipe', () => {
-  it('create an instance', () => {
-    const pipe = new ContainerTitlePipe();
-    expect(pipe).toBeTruthy();
-  });
-});
