+++ src/app/pipes/index.ts
@@ -5,5 +5,4 @@ export { EmbracePipe } from './embrace.pipe';
 export { ContainerPipe } from './container.pipe';
 export { IdentifierPipe } from './identifier.pipe';
 export { PrefixPipe } from './prefix.pipe';
-export { ContainerTitlePipe } from './container-title.pipe';
 export * from './type-pipes';
